Backend ( Django )/requirements.txt
asgiref==3.7.2
beautifulsoup4==4.11.2
bs4==0.0.1
certifi==2022.12.7
chardet==3.0.4
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
distlib==0.3.6
Django==4.2.6
djangorestframework==3.14.0
filelock==3.12.0
google-trans-new==1.1.9
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
joblib==1.2.0
nltk==3.8.1
numpy==1.24.3
pandas==2.0.1
pip==22.3.1
platformdirs==3.5.0
python-dateutil==2.8.2
pytz==2023.3
regex==2022.10.31
requests==2.28.2
rfc3986==1.5.0
setuptools==65.5.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
sqlparse==0.4.4
textblob==0.17.1
tqdm==4.64.1
tzdata==2023.3
urllib3==1.26.14
virtualenv==20.23.0
